Remove duplicate information about the Google Cloud Storage Loader
Currently, in our manual setup guides, we tend to point the reader to
pertinent information in the Pipeline Components and Applications section
at the first opportunity. For the GCS loader, we had duplicate information
in the Pipeline Components and Applications section and the GCP Manual Setup
section, so I have removed the duplicate info.