Link to RDB loader in relation to loading into Snowflake
In our manual setup guides, we signpost that it is possible to set up
loading into Snowflake. However, Snowflake was not mentioned in the
GCP manual setup, and the AWS setup still pointed to the (now
deprecated Snowplow Snowflake Loader).

So here we:
- ensure that we point to RDB loader for loading into Snowflake
for both the AWS and GCP setup
- consistently refer to Snowflake (rather than sometimes SnowflakeDB)
- add a deprecation warning to the docs for Snowplow Snowflake Loader