Move previous version RDB loader docs out of AWS manual setup
We have a previous versions section in the RDB loader docs. We had some of the docs
for previous versions living in the AWS manual setup, so we have moved them to the
previous versions section.

We also had very similar docs for the legacy EmrEtlRunner, so I have removed one of
the duplicates, and pointed to the docs in the legacy Pipeline Components section.

The RDB loader upgrade docs were living in specifically the RDB loader 2.x.x section,
so they are now moved one level up. We would like people on old versions to upgrade, so
these should be findable.