javascript-tracker-v3: more prominent custom build docs (#359)
Make documentation on building a version of sp.js v3 with custom plugins a bit more prominent. This is a small rewrite of the information located in the migration guide, but elevated to its own page (since it couldn't actually be done easily in v2 this location isn't an intuitive place to look).

I've also added some info on bundling in external plugins, which isn't covered on the old page.