Rename most usages of “bad rows” to “failed events”
“Failed events” and “bad rows” are used interchangeably in the docs,
however we should have a single term for this concept (which is “failed events”).

That said, it still makes sense to use “bad rows” in parts of the docs that
are closely tied to working with the (soon to be legacy) badrow format.

This commit changes all colloquial usages of “bad rows” to “failed events”
and retains “bad rows” in APIs and pages related to recovery and querying S3/GCS.

In the future, we might need to distinguish between “failed events, new format”
(warehouse) and “failed events, old format” (aka “bad rows”, S3/GCS), but we will
do so as needed by searching for “failed events” references and qualifying them.
In most contexts, only the fact that there is a failed event matters, not the format.